+++ demoapp/src/main/java/com/transferwise/kafka/tkms/demoapp/KafkaConfiguration.java
@@ -34,6 +34,7 @@ public class KafkaConfiguration implements InitializingBean {
   @Override
   public void afterPropertiesSet() {
     AdminClient adminClient = AdminClient.create(kafkaAdmin.getConfigurationProperties());
+
     try {
       createTopic(adminClient, new NewTopic("MyTopic", 10, (short) 1));
       createTopic(adminClient, new NewTopic("ComplexTest0", 10, (short) 1));