+++ tw-tkms-starter/src/main/java/com/transferwise/kafka/tkms/ITkmsTopicValidator.java
@@ -1,0 +1,10 @@
+package com.transferwise.kafka.tkms;
+
+import com.transferwise.kafka.tkms.api.TkmsShardPartition;
+
+public interface ITkmsTopicValidator {
+
+  void validate();
+
+  void validate(TkmsShardPartition tkmsShardPartition, String topic, Integer partition);
+}