+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.27.0] - 2023-12-20
+
+### Changed
+
+- Topic validation can now be done via Kafka Admin, instead of Kafka Producer.
+  The new logic is under a feature flag, until the logic gets more battle tested.
+
 ## [0.26.0] - 2023-12-14
 
 ### Removed